requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langchain-google-genai>=0.0.6
google-generativeai>=0.3.2
python-dotenv==1.0.0
pydantic>=2.5.3
chromadb>=0.4.22
tiktoken>=0.5.2
markdown>=3.5.2
flask>=3.0.0
gunicorn>=21.2.0 